pyproject.toml
[project]
name = "agent-communication"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "richard.hipsh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "poetry (>=2.1.4,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "redis[hiredis] (>=5.0.0,<6.0.0)",
    "aio-pika (>=9.4.0,<10.0.0)",
    "types-redis (>=4.6.0.20241004,<5.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
mypy = "^1.17.1"
pre-commit = "^4.3.0"
pytest-asyncio = "^0.23.0"
testcontainers = "^4.0.0"
pytest-docker = "^3.0.0"
pika = "^1.3.2"
types-redis = "^4.6.0.20241004"

[tool.mypy]
strict = true
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
ignore_missing_imports = true


[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
line-length = 88
target-version = "py39"
